pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decklister"
version = "1.0.0"
description = "Deck image generator for Star Wars Unlimited"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Pillow",
    "requests",
    "PySide6",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
decklister = "decklister.__main__:main_gui"
decklister-cli = "decklister.__main__:main_cli"

[tool.setuptools.packages.find]
include = ["decklister*"]

[tool.setuptools.package-data]
decklister = ["*.png", "*.ico"]